pyproject.toml
[tool.poetry]
name = "LAB_GUI"
version = "0.1.0"
description = "GUI for high voltage controller"
authors = ["Yunosuke Ohsugi <63651161+5yutan5@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/5yutan5/LAB_GUI"
homepage = "https://github.com/5yutan5/LAB_GUI"
packages = [{ include = "diy_hv", from = "DIYHV" }]

keywords = []

classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
pyqtdarktheme = "^1.0.3"
QtPy = "^2.0.0"
darkdetect = "^0.5.1"
QtAwesome = "^1.1.1"
pyserial = "^3.5"
rich = "^11.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-return = "^1.1.3"
flake8-print = "^4.0.0"
flake8-docstrings = "^1.6.0"
flake8-bugbear = "^22.1.11"
flake8-simplify = "^0.15.1"
flake8-comprehensions = "^3.8.0"
flake8-eradicate = "^1.2.0"
flake8-rst-docstrings = "^0.2.5"
flake8-pytest-style = "^1.6.0"
pep8-naming = "^0.12.1"
pyproject-flake8 = "^0.0.1-alpha.2"

pytest = "^6.2.5"
pytest-cov = { version = "^3.0.0", extras = ["toml"] }
pytest-xvfb = "^2.0.0"
pytest-randomly = "^3.11.0"
pytest-qt = "^4.0.2"
pytest-github-actions-annotate-failures = "^0.1.6"

black = "^21.12b0"
isort = "^5.10.1"
pre-commit = "^2.17.0"
pyupgrade = "^2.31.0"
pyinstaller = "^4.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."pyinstaller40"]
hook-dirs = "diy_hv.__pyinstaller:get_hook_dirs"

[tool.flake8]
max-line-length = 119
max-complexity = 10
docstring-convention = "google"
per-file-ignores = ["**/__init__.py:F401"]

[tool.black]
line-length = 119

[tool.isort]
profile = "black"
line_length = 119

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if __name__ == .__main__.:"]
